ref(onboarding): Split repo search and selection into separate hooks
Extract useRepoSearch (search query + dropdown items) and
useRepoSelection (add/remove lifecycle + existing repo lookup) from
scmRepoSelector, leaving it as pure UI.

Refs VDY-19
